+++ error/options_test.go
@@ -21,3 +21,20 @@ func TestHandler(t *testing.T) {
 
 	assert.Equal(t, handler, opts.handlers[0])
 }
+
+func TestConditional(t *testing.T) {
+	handler := new(emperror.TestHandler)
+	option := Handler(handler)
+
+	t.Run("condition met", func(t *testing.T) {
+		opts := newOptions(Conditional(true, option))
+
+		assert.Equal(t, handler, opts.handlers[0])
+	})
+
+	t.Run("condition not met", func(t *testing.T) {
+		opts := newOptions(Conditional(false, option))
+
+		assert.Nil(t, opts.handlers)
+	})
+}